pyproject.toml
[tool.poetry]
name = "somlier"
version = "0.6.2"
description = ""
authors = ["socar-hardy <hardy@socar.kr>"]
packages = [
    {include = "somlier"}
]

[tool.poetry.scripts]
somlier = "somlier.__main__:main"

[tool.poetry.dependencies]
python = ">=3.7.1,<3.9"
mlflow = "^1.21.0"
pydantic = "^1.8.1"
fire = "^0.4.0"
google-cloud-storage = "^1.38.0"
python-dotenv = "^0.17.1"
fastapi = {extras = ["all"], version = "^0.65.1"}
BentoML = "^0.13.0"
kubernetes = "^17.17.0"
loguru = "^0.5.3"
socar-data-dag-builder = {git = "git@github.com:socar-inc/socar-data-dag-builder.git", rev = "v0.4.3"}
GitPython = { version = "^3.1.24" }
dependency-injector = "^4.37.0"
pandas = "^1.2.0"
toml = "^0.10.2"
scipy = ">=1.7.0,<1.8.0"
MarkupSafe = "2.0.1"
ddtrace = "^1.1.3"
joblib = "^1.1.0"
scikit-learn = "1.0.2"
torch = "1.9.0"
protobuf = "3.19.0"
streamlit = "^1.12.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.12.1"
black = "^21.5b0"
pycln = "^1.0.3"
pytest = "^6.2.4"
PyMySQL = "^1.0.2"
cryptography = "^3.4.7"
commitizen = "^2.17.6"
isort = "^5.8.0"
locust = "^1.6.0"
pytest-dotenv = "^0.5.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.pycln]
all = true

[tool.isort]
profile = "black"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.6.2"
version_files = [
    "pyproject.toml:version",
    "somlier/__init__.py:__version__"
]
tag_format = "v$version"
update_changelog_on_bump = true
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]
